📛rename:  사용자 프로필 버튼 폴더 이동
- component -> component/button

사용자 프로필 컴포넌트 구현 #38
